+++ Sources/RichTextKit/_Deprecated/RichTextCommand+AlignmentOptionsGroup.swift
@@ -10,13 +10,7 @@ import SwiftUI
 
 public extension RichTextCommand {
 
-    /**
-     This view can add list of text alignment options to the
-     main menu, using an ``RichTextCommand/ActionButtonGroup``.
-     
-     This view requires that a ``RichTextContext`` is set as
-     a focused value, otherwise it will be disabled.
-     */
+    @available(*, deprecated, message: "Use RichTextCommand.ActionButtonGroup alignments initializer instead.")
     struct AlignmentOptionsGroup: View {
 
         public init() {}