+++ Sources/RichTextKit/_Deprecated/RichTextCommand+StyleOptionsGroup.swift
@@ -10,13 +10,7 @@ import SwiftUI
 
 public extension RichTextCommand {
 
-    /**
-     This view can add a list of styling options to the main
-     menu, using an ``RichTextCommand/ActionButtonGroup``.
-     
-     This view requires that a ``RichTextContext`` is set as
-     a focused value, otherwise it will be disabled.
-     */
+    @available(*, deprecated, message: "Use RichTextCommand.ActionButtonGroup styles initializer instead.")
     struct StyleOptionsGroup: View {
 
         public init() {}