+++ Sources/RichTextKit/_Deprecated/RichTextCommand+SuperscriptOptionsGroup.swift
@@ -10,13 +10,7 @@ import SwiftUI
 
 public extension RichTextCommand {
 
-    /**
-     This group can add a list of superscript options to the
-     main menu, using an ``RichTextCommand/ActionButtonGroup``.
-     
-     This view requires that a ``RichTextContext`` is set as
-     a focused value, otherwise it will be disabled.
-     */
+    @available(*, deprecated, message: "Use RichTextCommand.ActionButtonGroup superscript initializer instead.")
     struct SuperscriptOptionsGroup: View {
 
         public var body: some View {